# docs/swagger.yaml
basePath: /
definitions:
  dto.AllCollectionsDataResponseSwagger:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CollectionsResponseSwagger'
        type: array
    type: object
  dto.CollectionDataResponseSwagger:
    properties:
      data:
        $ref: '#/definitions/dto.CollectionsResponseSwagger'
    type: object
  dto.CollectionItemsDataResponseSwagger:
    properties:
      data:
        $ref: '#/definitions/dto.CollectionItemsResponseSwagger'
    type: object
  dto.CollectionItemsResponseSwagger:
    properties:
      collection:
        type: string
      copyNumber:
        items:
          type: string
        type: array
      description:
        type: string
      edition:
        type: string
      entities:
        items:
          $ref: '#/definitions/dto.EntitiesSwagger'
        type: array
      id:
        type: string
      images:
        items:
          type: string
        type: array
      isOwner:
        type: boolean
      itemType:
        $ref: '#/definitions/models.ItemTypes'
      name:
        type: string
      owner:
        $ref: '#/definitions/dto.UsersSwagger'
      platform:
        $ref: '#/definitions/models.Platforms'
      purchaseDate:
        type: string
      purchasePrice:
        type: number
      rating:
        type: number
      sealed:
        type: boolean
      shippingCost:
        type: number
    type: object
  dto.CollectionItemsSwagger:
    properties:
      collections:
        items:
          $ref: '#/definitions/dto.CollectionsSwagger'
        type: array
      copyNumber:
        items:
          type: string
        type: array
      createdAt:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      edition:
        type: string
      entities:
        items:
          $ref: '#/definitions/dto.EntitiesSwagger'
        type: array
      id:
        type: string
      images:
        items:
          type: string
        type: array
      itemType:
        $ref: '#/definitions/models.ItemTypes'
      itemTypeID:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/dto.UsersSwagger'
      platform:
        $ref: '#/definitions/models.Platforms'
      platformID:
        type: string
      purchaseDate:
        type: string
      purchasePrice:
        type: number
      rating:
        type: number
      sealed:
        type: boolean
      shippingCost:
        type: number
      transliteration:
        type: string
      updatedAt:
        type: string
      userLogin:
        type: string
    type: object
  dto.CollectionsResponseSwagger:
    properties:
      album:
        items:
          type: string
        type: array
      bannerUrl:
        type: string
      canLike:
        type: boolean
      canSubscribe:
        type: boolean
      collectionItems:
        items:
          $ref: '#/definitions/dto.CollectionItemsResponseSwagger'
        type: array
      collectionItemsCount:
        type: integer
      created:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      isOwner:
        type: boolean
      isPrivate:
        type: boolean
      likesCount:
        type: integer
      name:
        type: string
      shareString:
        type: string
      shippingTotal:
        type: number
      subscribersCount:
        type: integer
      tags:
        items:
          $ref: '#/definitions/dto.TagsSwagger'
        type: array
      totalPrice:
        type: number
      transliteration:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponseSwagger'
    type: object
  dto.CollectionsSwagger:
    properties:
      album:
        items:
          type: string
        type: array
      bannerUrl:
        type: string
      collectionItems:
        items:
          $ref: '#/definitions/dto.CollectionItemsSwagger'
        type: array
      created:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      isPrivate:
        type: boolean
      likes:
        items:
          type: string
        type: array
      name:
        type: string
      shareString:
        type: string
      shippingTotal:
        type: integer
      subscribersCount:
        type: integer
      tags:
        items:
          $ref: '#/definitions/dto.TagsSwagger'
        type: array
      totalPrice:
        type: integer
      transliteration:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/dto.UsersSwagger'
      userLogin:
        type: string
    type: object
...
  dto.CommonResponse:
    properties:
      data:
        $ref: '#/definitions/dto.Resp'
    type: object
...
  dto.EntitiesSwagger:
    properties:
      collectionItem:
        items:
          $ref: '#/definitions/dto.CollectionItemsSwagger'
        type: array
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.Resp:
    properties:
      success:
        type: boolean
    type: object
  dto.TagsSwagger:
    properties:
      collections:
        items:
          $ref: '#/definitions/dto.CollectionsSwagger'
        type: array
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.UserResponseSwagger:
    properties:
      avatarUrl:
        type: string
      backgroundUrl:
        type: string
      canSubscribe:
        type: boolean
      collectionItemsCount:
        default: 0
        type: integer
      collectionSubscriptions:
        items:
          type: string
        type: array
      collectionsCount:
        default: 0
        type: integer
      created:
        type: string
      dislikes:
        type: integer
      email:
        type: string
      likes:
        type: integer
      login:
        type: string
      subscribers:
        type: integer
      subscriptions:
        items:
          type: string
        type: array
      telegramId:
        type: string
      totalSum:
        default: 0
        type: integer
      userName:
        type: string
      vkId:
        type: string
    type: object
  dto.UsersSwagger:
    properties:
      avatarUrl:
        type: string
      backgroundUrl:
        type: string
      collectionSubscriptions:
        items:
          type: string
        type: array
      created:
        type: string
      dislikes:
        type: integer
      email:
        type: string
      likes:
        type: integer
      login:
        type: string
      subscribers:
        type: integer
      subscriptions:
        items:
          type: string
        type: array
      telegramId:
        type: string
      userName:
        type: string
      verificationToken:
        type: string
      vkId:
        type: string
    type: object
...
  models.CollectionCreateRequest:
    properties:
      bannerUrl:
        type: string
      description:
        type: string
      isPrivate:
        type: boolean
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      transliteration:
        type: string
      userLogin:
        type: string
    type: object
  models.CollectionItemsCopyOrMoveRequest:
    properties:
      id:
        type: string
      sourceCollectionId:
        type: string
      targetCollectionIds:
        items:
          type: string
        type: array
    type: object
...
  models.CollectionItemsRequestCreate:
    properties:
      collection:
        type: string
      copyNumber:
        items:
          type: integer
        type: array
      description:
        type: string
      edition:
        type: string
      images:
        items:
          type: string
        type: array
      itemType:
        type: string
      name:
        type: string
      platform:
        type: string
      purchaseDate:
        type: string
      purchasePrice:
        type: number
      rating:
        type: number
      sealed:
        type: boolean
      shippingCost:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
...
  models.ItemTypes:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      ruName:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.Platforms:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
host: dev.lootor.me
info:
  contact: {}
  description: Документация к API Lootor.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lootor
  version: "1.0"
paths:
...
  /public/collection_item:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionItemsDataResponseSwagger'
      summary: получение одного КИ
      tags:
      - collection items
  /public/collection_item/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: search
        in: query
        name: search
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AllCollectionsDataResponseSwagger'
      summary: получить КИ all
      tags:
      - collection items
  /public/collections:
    get:
      consumes:
      - application/json
      parameters:
      - description: логин
        in: query
        name: userLogin
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: translit
        in: query
        name: transliteration
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      - description: поле сортировки% может быть  name, purchaseDate, purchasePrice,
          platform, type, rating
        in: query
        name: orderBy
        type: string
      - description: порядок сортировки asc или desc
        in: query
        name: order
        type: string
      - description: поиск по именам КИ внутри коллекции
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionDataResponseSwagger'
      summary: получение коллекций по id или по логину + транслит
      tags:
      - collections
  /public/collections/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: логин
        in: query
        name: userLogin
        required: true
        type: string
      - description: поле сортировки% может быть totalPrice, name, created, collectionItemsCount
        in: query
        name: orderBy
        type: string
      - description: порядок сортировки asc или desc
        in: query
        name: order
        type: string
      - description: поиск по именам коллекций
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AllCollectionsDataResponseSwagger'
      summary: получение коллекций по логину
      tags:
      - collections
  /public/collections/catalog:
    get:
      consumes:
      - application/json
      parameters:
      - description: поле сортировки% может быть totalPrice, name, created, collectionItemsCount,
          likesCount
        in: query
        name: orderBy
        type: string
      - description: порядок сортировки asc или desc
        in: query
        name: order
        type: string
      - description: поиск по именам коллекций
        in: query
        name: search
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AllCollectionsDataResponseSwagger'
      summary: получение коллекций всех
      tags:
      - collections
...
  /secured/collection_item:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: удаление КИ
      tags:
      - collection items
    post:
      consumes:
      - application/json
      parameters:
      - description: поля создания
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CollectionItemsRequestCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionItemsDataResponseSwagger'
      summary: Создание КИ
      tags:
      - collection items
  /secured/collection_item/copy:
    post:
      consumes:
      - application/json
      parameters:
      - description: polya
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CollectionItemsCopyOrMoveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: копирование/перемещение КИ
      tags:
      - collection items
  /secured/collection_item/like:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: like/dislike
      tags:
      - collection items
  /secured/collection_item/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: поля создания
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CollectionItemsRequestCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionItemsDataResponseSwagger'
      summary: обновление КИ
      tags:
      - collection items
  /secured/collections:
    post:
      consumes:
      - application/json
      parameters:
      - description: поля создания
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CollectionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionDataResponseSwagger'
      summary: Создание коллекции
      tags:
      - collections
  /secured/collections/like:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: like/dislike
      tags:
      - collections
  /secured/collections/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: поля создания
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/models.CollectionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CollectionDataResponseSwagger'
      summary: обновление коллекции
      tags:
      - collections
...
schemes:
- https
swagger: "2.0"